# polyval: ARMv8 `PMULL` backend