--- a/ghash/Cargo.toml
+++ b/ghash/Cargo.toml
@@ -24,4 +24,5 @@ hex-literal = "0.2"
 
 [features]
 std = ["polyval/std"]
+armv8 = ["polyval/armv8"]
 force-soft = ["polyval/force-soft"]