--- a/polyval/Cargo.toml
+++ b/polyval/Cargo.toml
@@ -15,19 +15,21 @@ categories = ["cryptography", "no-std"]
 edition = "2018"
 
 [dependencies]
+cfg-if = "1"
 opaque-debug = "0.3"
 universal-hash = { version = "0.4", default-features = false }
 zeroize = { version = "1.3", optional = true, default-features = false }
 
-[target.'cfg(any(target_arch = "x86_64", target_arch = "x86"))'.dependencies]
-cpufeatures = "0.1"
+[target.'cfg(any(target_arch = "aarch64", target_arch = "x86_64", target_arch = "x86"))'.dependencies]
+cpufeatures = "0.1.4"
 
 [dev-dependencies]
 hex-literal = "0.2"
 
 [features]
-force-soft = [] # Disable support for hardware intrinsics (CLMUL)
 std = ["universal-hash/std"]
+armv8 = [] # Enable nightly-only ARMv8 intrinsics support
+force-soft = [] # Disable support for hardware intrinsics
 
 [package.metadata.docs.rs]
 all-features = true