--- a/polyval/src/backend.rs
+++ b/polyval/src/backend.rs
@@ -1,29 +1,24 @@
 //! POLYVAL backends
 
-#[cfg(all(
-    any(target_arch = "x86", target_arch = "x86_64"),
-    not(feature = "force-soft")
-))]
-pub(crate) mod autodetect;
-
-#[cfg(all(
-    any(target_arch = "x86", target_arch = "x86_64"),
-    not(feature = "force-soft")
-))]
-pub(crate) mod clmul;
-
 #[cfg_attr(not(target_pointer_width = "64"), path = "backend/soft32.rs")]
 #[cfg_attr(target_pointer_width = "64", path = "backend/soft64.rs")]
-pub(crate) mod soft;
+mod soft;
 
-#[cfg(all(
-    any(target_arch = "x86", target_arch = "x86_64"),
-    not(feature = "force-soft")
-))]
-pub use crate::backend::autodetect::Polyval;
+use cfg_if::cfg_if;
 
-#[cfg(not(all(
-    any(target_arch = "x86", target_arch = "x86_64"),
-    not(feature = "force-soft")
-)))]
-pub use crate::backend::soft::Polyval;
+cfg_if! {
+    if #[cfg(all(target_arch = "aarch64", feature = "armv8", not(feature = "force-soft")))] {
+        mod autodetect;
+        mod pmull;
+        pub use crate::backend::autodetect::Polyval;
+    } else if #[cfg(all(
+        any(target_arch = "x86_64", target_arch = "x86"),
+        not(feature = "force-soft")
+    ))] {
+        mod autodetect;
+        mod clmul;
+        pub use crate::backend::autodetect::Polyval;
+    } else {
+        pub use crate::backend::soft::Polyval;
+    }
+}