@tarcieri 2021-05-31
This is using some fairly bleeding edge features for `stdsimd`.

For example, `vmull_p64` is only available in very recent nightly versions, to the point the docker image used by `cross` does not have a recent enough nightly to support it.