@tarcieri 2021-05-31
This is confirmed tested and working on an Apple M1, although unfortunately until the Docker container cross uses is updated, it unfortunately can't be tested in CI.

The main thing missing right now (besides some more extensive testing) is runtime detection for `PMULL` (i.e. adapting the existing `autodetect::Polyval` type to support autodetection on ARM64)

Preliminary benchmarks on an Apple M1:

## Before

```
test bench1_10    ... bench:          34 ns/iter (+/- 0) = 294 MB/s
test bench2_100   ... bench:         240 ns/iter (+/- 1) = 416 MB/s
test bench3_1000  ... bench:       2,166 ns/iter (+/- 10) = 461 MB/s
test bench3_10000 ... bench:      21,580 ns/iter (+/- 141) = 463 MB/s
```

## After

```
test bench1_10    ... bench:          16 ns/iter (+/- 9) = 625 MB/s
test bench2_100   ... bench:          78 ns/iter (+/- 0) = 1282 MB/s
test bench3_1000  ... bench:         663 ns/iter (+/- 1) = 1508 MB/s
test bench3_10000 ... bench:       6,573 ns/iter (+/- 67) = 1521 MB/s
```

An improvement, but still seems kind of slow. Perhaps we need better parallel APIs for `universal-hash`.